+++ seeyousun-backend/review-service/src/main/java/com/taass/seeyousun/reviewservice/messaging/ReviewMessageDTO.java
@@ -5,11 +5,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class ReviewMessageDTO {
     private Long resortId;
-    private Double averageRating;
+    private BigDecimal averageRating;
 }
